refactor(Collapse): rename to ts
