+++ components/collapse/style.ts
