+++ components/collapse/types.ts
